move projects into component
