+++ src/pages/projects/[[...slug]].js
@@ -1,4 +1,6 @@
 import { data } from '../../_data'
+import { Project } from '../../components/projects/Project'
+import { useRouter } from 'next/router'
 
 export default function Projects() {
   const router = useRouter()
@@ -46,7 +48,6 @@ export const Explorer = props => {
     return (
       <Stack spacing="6" py="5" px="8" divider={<StackDivider />}>
         {repos?.map(repo => {
-          console.log(repo)
           return (
             <Repo
               org={repo.org}
@@ -79,50 +80,8 @@ export const Explorer = props => {
   )
 }
 
-import { Box, HStack, IconButton } from '@chakra-ui/react'
-import { HiCollection, HiPencilAlt, HiTrash } from 'react-icons/hi'
-import { useRouter } from 'next/router'
-import plur from 'plur'
-
-export const Project = props => {
-  const router = useRouter()
-  const { title, children, repos, href } = props
-
-  return (
-    <Box position="relative" as="a" href={`${router.asPath}/${href}`}>
-      <Box fontWeight="bold" maxW="xl">
-        {title}
-      </Box>
-      <HStack fontSize="sm" fontWeight="medium" color={mode('gray.500', 'white')} mt="1">
-        <Box as={HiCollection} fontSize="md" color="gray.400" />
-        <span>{repos} {plur('repository', repos)}</span>
-      </HStack>
-      <Box mt="3" maxW="xl" color={mode('gray.600', 'gray.200')}>
-        {children}
-      </Box>
-      <HStack
-        position={{
-          sm: 'absolute',
-        }}
-        top={{
-          sm: '0',
-        }}
-        insetEnd={{
-          sm: '0',
-        }}
-        mt={{
-          base: '4',
-          sm: '0',
-        }}
-      >
-        <IconButton aria-label="Edit" icon={<HiPencilAlt />} rounded="full" size="sm" />
-        <IconButton aria-label="Delete" icon={<HiTrash />} rounded="full" size="sm" />
-      </HStack>
-    </Box>
-  )
-}
-
 import { FaGithubAlt } from 'react-icons/fa'
+import { BsFillFileEarmarkBinaryFill } from 'react-icons/bs'
 
 export const Repo = props => {
   const { repo, org, children, href } = props
@@ -134,7 +93,10 @@ export const Repo = props => {
         {repo}
       </Box>
       <HStack fontSize="sm" fontWeight="medium" color={mode('gray.500', 'white')} mt="1">
-        <Box as={HiCollection} fontSize="md" color="gray.400" />
+        <Box as={BsFillFileEarmarkBinaryFill} fontSize="md" color="gray.400" />
+        <Text>
+          7b
+        </Text>
       </HStack>
       <Box mt="3" maxW="xl" color={mode('gray.600', 'gray.200')}>
         {children}
@@ -163,8 +125,8 @@ export const Repo = props => {
 }
 
 import {
-  Button,
-  Divider,
+  Button, Box, HStack,
+  Divider, IconButton,
   Flex,
   Stack,
   StackDivider,
